refactor distance functions to utils.rs
